Move ProjectSelectionManager.h to lib-time-frequency-selection
