+++ libraries/lib-time-frequency-selection/ProjectSelectionManager.cpp
